bug: RecentSearchView 분리 (#68)
