+++ DietLog-MVVM/View/Serach/View/Cells/RecentWordCollectionViewCell.swift
